fixes admin view (download)
#41305: Admintoolbar can be activated without rights or login (Ilja Lukin (Fachhochschule Dortmund))

Cherry-picked and adapted from
https://github.com/ILIAS-eLearning/ILIAS/commit/0400626c3805955837ae3cadfa60877262378312

+++ Services/Container/classes/class.ilContainerGUI.php
@@ -1635,7 +1635,7 @@ class ilContainerGUI extends ilObjectGUI implements ilDesktopItemHandling
     
     public function downloadObject()
     {
-        if (in_array($this->user->getId(), [ANONYMOUS_USER_ID, 0], true)) {
+        if (in_array($this->user->getId(), [(int) ANONYMOUS_USER_ID, 0], true)) {
             return;
         }
